+++ Mtf/Client/Driver/Selenium/WaitUntil.php
@@ -16,58 +16,44 @@ namespace Mtf\Client\Driver\Selenium;
  */
 class WaitUntil
 {
-    /**
-     * PHPUnit Test Case instance
-     *
-     * @var TestCase
-     */
-    private $testCase;
 
     /**
      * Default timeout, ms
      *
      * @var int
      */
-    private $defaultTimeout = 40000;
+    const DEFAULT_TIMEOUT = 40000;
 
     /**
      * The sleep interval between iterations, ms
      *
      * @var int
      */
-    private $defaultSleepInterval = 500;
-
-    /**
-     * @constructor
-     * @param TestCase $testCase
-     */
-    public function __construct(TestCase $testCase)
-    {
-        $this->testCase = $testCase;
-    }
+    const DEFAULT_SLEEP_INTERVAL = 500;
 
     /**
      * Run timeout waiting script
      *
      * @param callback $callback Callback to run until it returns not null or timeout occurs
+     * @param $testCase
      * @param null|int $timeout
      * @return mixed
      * @throws \PHPUnit_Extensions_Selenium2TestCase_Exception
      * @throws \PHPUnit_Extensions_Selenium2TestCase_WebDriverException
      */
-    public function run($callback, $timeout = null)
+    public static function run($callback, $testCase, $timeout = null)
     {
         if (!is_callable($callback)) {
             throw new \PHPUnit_Extensions_Selenium2TestCase_Exception('The valid callback is expected');
         }
 
         // if there was an implicit timeout specified - remember it and temporarily turn it off
-        $implicitWait = $this->testCase->timeouts()->getLastImplicitWaitValue();
+        $implicitWait = $testCase->timeouts()->getLastImplicitWaitValue();
         if ($implicitWait) {
-            $this->testCase->timeouts()->implicitWait(0);
+            $testCase->timeouts()->implicitWait(0);
         }
         if (is_null($timeout)) {
-            $timeout = $this->defaultTimeout;
+            $timeout = self::DEFAULT_TIMEOUT;
         }
         $timeout /= 1000;
         $endTime = microtime(true) + $timeout;
@@ -77,7 +63,7 @@ class WaitUntil
                 $result = call_user_func($callback);
                 if (!is_null($result)) {
                     if ($implicitWait) {
-                        $this->testCase->timeouts()->implicitWait($implicitWait);
+                        $testCase->timeouts()->implicitWait($implicitWait);
                     }
                     return $result;
                 }
@@ -86,7 +72,7 @@ class WaitUntil
             }
             if (microtime(true) > $endTime) {
                 if ($implicitWait) {
-                    $this->testCase->timeouts()->implicitWait($implicitWait);
+                    $testCase->timeouts()->implicitWait($implicitWait);
                 }
                 $message = "Timed out after {$timeout} second" . ($timeout != 1 ? 's' : '');
                 throw new \PHPUnit_Extensions_Selenium2TestCase_WebDriverException(
@@ -95,7 +81,7 @@ class WaitUntil
                     $lastException
                 );
             }
-            usleep($this->defaultSleepInterval * 1000);
+            usleep(self::DEFAULT_SLEEP_INTERVAL * 1000);
         }
     }
 }