+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+* Added logic to get a list of enabled modules in target Magento instance
+* Changed module file resolver to only search files in modules that are enabled in target Magento instance
+* Added dynamic generation of page classes based on enabled modules in target Magento instance
 * Changed path to generated page classes
 * Implemented recursive re-initialization for Block
 * Implemented scenario configuration reader
 * Implemented recursive merge for injectable fixture data from repository and variations(*.csv file)
+* Updated scenario configuration reader and fixed issues with preparing step sequence
 
 1.0.0-dev.6
 =============