+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed random wait issues in parallel mode
 * Added logic to run a subset of tests based on affected modules
 * Added logic to get a list of enabled modules in target Magento instance
 * Changed module file resolver to only search files in modules that are enabled in target Magento instance